Human Gut Microbiome Abundance and Patient Classes

Data set consists of a matrix of abundances of 1192 microbial taxa for 575 samples and a factor vector of classes for Parkinson's disease for the same 575 patients

Format

x has a row for each sample and a column for each taxon. lubomPD is a factor vector with values 0 representing Healthy Control (HC) and 1 representing Parkinson's Disease (PD).

Source

An R package PD16Sdata, BMC, 2023. Webpage: https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-023-01475-4